refactor: extract types to separate files
